Add options panel to manage chat room auto-joins.
Provides a listing of available chat rooms and allows users to select
which rooms will be auto-joined when connecting to chat.

On first time connections, the default rooms will be added to the
auto-join list, and then are managed through the auto-join settings as
any other room.

Room list is sorted alphabetically, ascending.

Changing an auto-join setting will not trigger an immediate join or
leave. Settings are applied the next time chat is connected.